♻️(schema-bench): refactor executor utilities into shared modules
- Extract common CLI utilities into separate files for better maintainability
  - parseArgs: Command-line argument parsing for dataset selection
  - selectTargetDatasets: Dataset selection logic (supports -all, specific datasets)
  - filterAndResolveDatasets: Path validation and resolution
  - loadInputFiles: Generic input file loading with schema validation
  - saveOutputFile: Output file management without archival
  - discoverDefaultDatasets/listAllDatasets: Dataset discovery utilities

- Refactor executeLiamDbShared.ts and executeOpenaiUnified.ts to use shared utilities
- Remove duplicate code between LiamDB and OpenAI executors
- Simplify workspace utility by removing unused getWorkspaceSubPath function
- Improve code organization and reusability across executor implementations
