+++ frontend/internal-packages/schema-bench/src/cli/executeLiamDbShared.ts
@@ -1,20 +1,15 @@
 #!/usr/bin/env node
 
-import { existsSync, mkdirSync } from 'node:fs'
-import { readdir, readFile, writeFile } from 'node:fs/promises'
-import { join, resolve } from 'node:path'
+import { resolve } from 'node:path'
 import { config } from 'dotenv'
-import {
-  err,
-  fromPromise,
-  ok,
-  type Result,
-  Result as ResultClass,
-} from 'neverthrow'
+import { err, ok, type Result } from 'neverthrow'
 import * as v from 'valibot'
 import { execute, type LiamDbExecutorInput } from '../executors/liamDb/index.ts'
+import { loadInputFiles, saveOutputFile } from './utils'
 
 config({ path: resolve(__dirname, '../../../../../.env') })
+
+// Outputs overwrite the latest files (no per-run archiving)
 
 const InputSchema = v.union([
   v.object({
@@ -29,103 +24,9 @@ type DatasetResult = {
   failure: number
 }
 
-async function loadInputFiles(
-  datasetPath: string,
-): Promise<
-  Result<Array<{ caseId: string; input: LiamDbExecutorInput }>, Error>
-> {
-  const inputDir = join(datasetPath, 'execution/input')
+// loadInputFiles moved to ./utils and used with schema + normalizer
 
-  if (!existsSync(inputDir)) {
-    return err(
-      new Error(
-        `Input directory not found: ${inputDir}. Please run setup-workspace first.`,
-      ),
-    )
-  }
-
-  const filesResult = await fromPromise(readdir(inputDir), (error) =>
-    error instanceof Error ? error : new Error('Failed to read directory'),
-  )
-
-  if (filesResult.isErr()) {
-    return err(filesResult.error)
-  }
-
-  const jsonFiles = filesResult.value.filter((file) => file.endsWith('.json'))
-  const inputs: Array<{ caseId: string; input: LiamDbExecutorInput }> = []
-
-  for (const file of jsonFiles) {
-    const caseId = file.replace('.json', '')
-    const contentResult = await fromPromise(
-      readFile(join(inputDir, file), 'utf-8'),
-      (error) =>
-        error instanceof Error
-          ? error
-          : new Error(`Failed to read file ${file}`),
-    )
-
-    if (contentResult.isErr()) {
-      return err(contentResult.error)
-    }
-
-    const parseResult = ResultClass.fromThrowable(
-      () => JSON.parse(contentResult.value),
-      (error) =>
-        error instanceof Error
-          ? error
-          : new Error(`Failed to parse JSON in ${file}`),
-    )()
-
-    if (parseResult.isErr()) {
-      return err(parseResult.error)
-    }
-
-    const validationResult = v.safeParse(InputSchema, parseResult.value)
-    if (!validationResult.success) {
-      return err(
-        new Error(
-          `Invalid input format in ${file}: ${JSON.stringify(validationResult.issues)}`,
-        ),
-      )
-    }
-
-    const normalizedInput: LiamDbExecutorInput =
-      typeof validationResult.output === 'string'
-        ? { input: validationResult.output }
-        : validationResult.output
-
-    inputs.push({
-      caseId,
-      input: normalizedInput,
-    })
-  }
-
-  return ok(inputs)
-}
-
-async function saveOutputFile(
-  datasetPath: string,
-  caseId: string,
-  output: unknown,
-): Promise<Result<void, Error>> {
-  const outputDir = join(datasetPath, 'execution/output')
-
-  if (!existsSync(outputDir)) {
-    mkdirSync(outputDir, { recursive: true })
-  }
-
-  const outputPath = join(outputDir, `${caseId}.json`)
-  const writeResult = await fromPromise(
-    writeFile(outputPath, JSON.stringify(output, null, 2)),
-    (error) =>
-      error instanceof Error
-        ? error
-        : new Error(`Failed to save output for ${caseId}`),
-  )
-
-  return writeResult.map(() => undefined)
-}
+// saveOutputFile moved to ./utils with optional archive
 
 async function executeCase(
   datasetPath: string,
@@ -151,7 +52,14 @@ export async function processDataset(
   datasetPath: string,
 ): Promise<DatasetResult> {
   // Load input files
-  const inputsResult = await loadInputFiles(datasetPath)
+  const inputsResult = await loadInputFiles<
+    typeof InputSchema,
+    LiamDbExecutorInput
+  >(datasetPath, InputSchema, (value) =>
+    typeof value === 'string'
+      ? { input: value }
+      : (value as LiamDbExecutorInput),
+  )
   if (inputsResult.isErr()) {
     console.warn(`⚠️  ${datasetName}: ${inputsResult.error.message}`)
     return { datasetName, success: 0, failure: 1 }
