+++ frontend/internal-packages/schema-bench/src/cli/executeLiamDbUnified.ts
@@ -1,101 +1,23 @@
 #!/usr/bin/env node
 
-import { existsSync, readdirSync } from 'node:fs'
-import { isAbsolute, join, relative, resolve } from 'node:path'
+// path resolution handled in utils
 import { processDataset } from './executeLiamDbShared'
 import {
+  filterAndResolveDatasets,
   getWorkspacePath,
   handleCliError,
   handleUnexpectedError,
+  parseArgs,
+  selectTargetDatasets,
 } from './utils'
 
-type CliOptions = {
-  datasets?: string[]
-  useAll?: boolean
-}
+// parseArgs moved to ./utils
 
-// Allow alphanumerics, '-', '_', 1..64 chars
-const VALID_DATASET = /^[A-Za-z0-9][A-Za-z0-9_-]{0,63}$/
-
-const parseArgs = (argv: string[]): CliOptions => {
-  // Supported patterns:
-  // - executeLiamDB -all
-  // - executeLiamDB -default -entity-extraction
-  const options: CliOptions = {}
-  const args = argv.slice(2)
-  for (const tok of args) {
-    if (!tok || !tok.startsWith('-')) continue
-    const name = tok.replace(/^-+/, '')
-    if (name === 'all') {
-      options.useAll = true
-      continue
-    }
-    if (name) {
-      if (!VALID_DATASET.test(name)) {
-        handleCliError(
-          `Invalid dataset token "${name}". Allowed: letters, numbers, "-", "_" (1–64 chars).`,
-        )
-      }
-      options.datasets = [...(options.datasets ?? []), name]
-    }
-  }
-  return options
-}
-
-const discoverDefaultDatasets = (workspacePath: string): string[] => {
-  // Default behavior: run only the 'default' dataset if it exists
-  const name = 'default'
-  return existsSync(join(workspacePath, name)) ? [name] : []
-}
-
-const listAllDatasets = (workspacePath: string): string[] => {
-  try {
-    const entries = readdirSync(workspacePath, { withFileTypes: true })
-    return entries.filter((e) => e.isDirectory()).map((e) => e.name)
-  } catch {
-    return []
-  }
-}
 type DatasetResult = { datasetName: string; success: number; failure: number }
 
-const selectTargetDatasets = (
-  options: CliOptions,
-  workspacePath: string,
-): string[] => {
-  let targets: string[] = []
-  if (options.useAll) {
-    targets = listAllDatasets(workspacePath)
-  } else if (!options.datasets || options.datasets.length === 0) {
-    targets = discoverDefaultDatasets(workspacePath)
-  }
-  if (options.datasets && options.datasets.length > 0) {
-    const union = new Set([...targets, ...options.datasets])
-    targets = Array.from(union)
-  }
-  return targets
-}
+// selectTargetDatasets moved to ./utils
 
-const filterAndResolveDatasets = (
-  targetDatasets: string[],
-  workspacePath: string,
-): Array<{ name: string; path: string }> => {
-  const resolvedWorkspace = resolve(workspacePath)
-  const resolved: Array<{ name: string; path: string }> = []
-  for (const name of targetDatasets) {
-    const datasetPath = resolve(workspacePath, name)
-    const rel = relative(resolvedWorkspace, datasetPath)
-    if (rel.startsWith('..') || isAbsolute(rel)) {
-      console.warn(`   ⚠️  Skipping invalid dataset path: ${name}`)
-      continue
-    }
-    if (!existsSync(datasetPath)) {
-      // Silently skip missing datasets to mirror existing behavior
-      continue
-    }
-    resolved.push({ name, path: datasetPath })
-  }
-  return resolved
-}
+// filterAndResolveDatasets moved to ./utils
 
 const runDatasets = async (
   datasets: Array<{ name: string; path: string }>,
