+++ frontend/internal-packages/schema-bench/src/cli/executeOpenaiUnified.ts
@@ -1,72 +1,23 @@
 #!/usr/bin/env node
 
-import { existsSync, mkdirSync, readdirSync } from 'node:fs'
-import { readdir, readFile, writeFile } from 'node:fs/promises'
-import { isAbsolute, join, relative, resolve } from 'node:path'
+import { resolve } from 'node:path'
 import { config as loadEnv } from 'dotenv'
-import {
-  err,
-  fromPromise,
-  ok,
-  type Result,
-  Result as ResultClass,
-} from 'neverthrow'
+import { err, ok, type Result } from 'neverthrow'
 import * as v from 'valibot'
 import { OpenAIExecutor } from '../executors/openai/openaiExecutor'
 import type { OpenAIExecutorInput } from '../executors/openai/types'
 import {
+  filterAndResolveDatasets,
   getWorkspacePath,
   handleCliError,
   handleUnexpectedError,
+  loadInputFiles,
+  parseArgs,
+  saveOutputFile,
+  selectTargetDatasets,
 } from './utils'
 
-type CliOptions = {
-  datasets?: string[]
-  useAll?: boolean
-}
-
-// Allow alphanumerics, '-', '_', 1..64 chars
-const VALID_DATASET = /^[A-Za-z0-9][A-Za-z0-9_-]{0,63}$/
-
-const parseArgs = (argv: string[]): CliOptions => {
-  // Supported patterns:
-  // - executeOpenai -all
-  // - executeOpenai -default -entity-extraction
-  const options: CliOptions = {}
-  const args = argv.slice(2)
-  for (const tok of args) {
-    if (!tok || !tok.startsWith('-')) continue
-    const name = tok.replace(/^-+/, '')
-    if (name === 'all') {
-      options.useAll = true
-      continue
-    }
-    if (name) {
-      if (!VALID_DATASET.test(name)) {
-        handleCliError(
-          `Invalid dataset token "${name}". Allowed: letters, numbers, "-", "_" (1–64 chars).`,
-        )
-      }
-      options.datasets = [...(options.datasets ?? []), name]
-    }
-  }
-  return options
-}
-
-const discoverDefaultDatasets = (workspacePath: string): string[] => {
-  // Default behavior: run only the 'default' dataset if it exists
-  const name = 'default'
-  return existsSync(join(workspacePath, name)) ? [name] : []
-}
-
-const listAllDatasets = (workspacePath: string): string[] => {
-  try {
-    const entries = readdirSync(workspacePath, { withFileTypes: true })
-    return entries.filter((e) => e.isDirectory()).map((e) => e.name)
-  } catch {
-    return []
-  }
-}
+// parseArgs moved to ./utils
 
 type DatasetResult = { datasetName: string; success: number; failure: number }
 
@@ -74,96 +25,9 @@ const InputSchema = v.object({
   input: v.string(),
 })
 
-async function loadInputFiles(
-  datasetPath: string,
-): Promise<
-  Result<Array<{ caseId: string; input: OpenAIExecutorInput }>, Error>
-> {
-  const inputDir = join(datasetPath, 'execution/input')
+// loadInputFiles moved to ./utils and used with schema + normalizer
 
-  if (!existsSync(inputDir)) {
-    return err(
-      new Error(
-        `Input directory not found: ${inputDir}. Please run setup-workspace first.`,
-      ),
-    )
-  }
-
-  const filesResult = await fromPromise(readdir(inputDir), (error) =>
-    error instanceof Error ? error : new Error('Failed to read directory'),
-  )
-
-  if (filesResult.isErr()) {
-    return err(filesResult.error)
-  }
-
-  const jsonFiles = filesResult.value.filter((file) => file.endsWith('.json'))
-  const inputs: Array<{ caseId: string; input: OpenAIExecutorInput }> = []
-
-  for (const file of jsonFiles) {
-    const caseId = file.replace('.json', '')
-    const contentResult = await fromPromise(
-      readFile(join(inputDir, file), 'utf-8'),
-      (error) =>
-        error instanceof Error
-          ? error
-          : new Error(`Failed to read file ${file}`),
-    )
-
-    if (contentResult.isErr()) {
-      return err(contentResult.error)
-    }
-
-    const parseResult = ResultClass.fromThrowable(
-      () => JSON.parse(contentResult.value),
-      (error) =>
-        error instanceof Error
-          ? error
-          : new Error(`Failed to parse JSON in ${file}`),
-    )()
-
-    if (parseResult.isErr()) {
-      return err(parseResult.error)
-    }
-
-    const validationResult = v.safeParse(InputSchema, parseResult.value)
-    if (!validationResult.success) {
-      return err(
-        new Error(
-          `Invalid input format in ${file}: ${JSON.stringify(validationResult.issues)}`,
-        ),
-      )
-    }
-
-    inputs.push({ caseId, input: validationResult.output })
-  }
-
-  return ok(inputs)
-}
-
-async function saveOutputFile(
-  datasetPath: string,
-  caseId: string,
-  output: unknown,
-): Promise<Result<void, Error>> {
-  const outputDir = join(datasetPath, 'execution/output')
-
-  if (!existsSync(outputDir)) {
-    mkdirSync(outputDir, { recursive: true })
-  }
-
-  // Write/refresh the latest flat file only (no per-run archiving)
-  const latestPath = join(outputDir, `${caseId}.json`)
-  const latestResult = await fromPromise(
-    writeFile(latestPath, JSON.stringify(output, null, 2)),
-    (error) =>
-      error instanceof Error
-        ? error
-        : new Error(`Failed to write latest output for ${caseId}`),
-  )
-
-  return latestResult.map(() => undefined)
-}
+// saveOutputFile moved to ./utils
 
 async function executeCase(
   executor: OpenAIExecutor,
@@ -185,51 +49,19 @@ async function executeCase(
   return ok(undefined)
 }
 
-const selectTargetDatasets = (
-  options: CliOptions,
-  workspacePath: string,
-): string[] => {
-  let targets: string[] = []
-  if (options.useAll) {
-    targets = listAllDatasets(workspacePath)
-  } else if (!options.datasets || options.datasets.length === 0) {
-    targets = discoverDefaultDatasets(workspacePath)
-  }
-  if (options.datasets && options.datasets.length > 0) {
-    const union = new Set([...targets, ...options.datasets])
-    targets = Array.from(union)
-  }
-  return targets
-}
+// selectTargetDatasets moved to ./utils
 
-const filterAndResolveDatasets = (
-  targetDatasets: string[],
-  workspacePath: string,
-): Array<{ name: string; path: string }> => {
-  const resolvedWorkspace = resolve(workspacePath)
-  const resolved: Array<{ name: string; path: string }> = []
-  for (const name of targetDatasets) {
-    const datasetPath = resolve(workspacePath, name)
-    const rel = relative(resolvedWorkspace, datasetPath)
-    if (rel.startsWith('..') || isAbsolute(rel)) {
-      console.warn(`   ⚠️  Skipping invalid dataset path: ${name}`)
-      continue
-    }
-    if (!existsSync(datasetPath)) {
-      // Silently skip missing datasets to mirror existing behavior
-      continue
-    }
-    resolved.push({ name, path: datasetPath })
-  }
-  return resolved
-}
+// filterAndResolveDatasets moved to ./utils
 
 async function processDataset(
   executor: OpenAIExecutor,
   datasetName: string,
   datasetPath: string,
 ): Promise<DatasetResult> {
-  const inputsResult = await loadInputFiles(datasetPath)
+  const inputsResult = await loadInputFiles<
+    typeof InputSchema,
+    OpenAIExecutorInput
+  >(datasetPath, InputSchema, (value) => value as OpenAIExecutorInput)
   if (inputsResult.isErr()) {
     console.warn(`⚠️  ${datasetName}: ${inputsResult.error.message}`)
     return { datasetName, success: 0, failure: 1 }
