+++ frontend/internal-packages/schema-bench/src/cli/utils/index.ts
@@ -1,2 +1,9 @@
+export * from './discoverDefaultDatasets.ts'
 export * from './error.ts'
+export * from './filterAndResolveDatasets.ts'
+export * from './listAllDatasets.ts'
+export * from './loadInputFiles.ts'
+export * from './parseArgs.ts'
+export * from './saveOutputFile.ts'
+export * from './selectTargetDatasets.ts'
 export * from './workspace.ts'
